Renamed the old dialog cs file to "Dialog [Obsolete].cs" and the new one to "Dialog.cs"
